Extract out contract types
