+++ app/src/helper/contracts.ts
@@ -1,0 +1,12 @@
+export type Governance = any;
+export type ArtifactApplication = any;
+export type ArtifactRegistry = any;
+export type Artists = any;
+export type Ens = any;
+
+export interface ContractListType {
+  Governance: Governance;
+  ArtifactApplication: ArtifactApplication;
+  ArtifactRegistry: ArtifactRegistry;
+  Artists: Artists;
+}
